requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
Babel==2.7.0
beautifulsoup4==4.6.0
billiard==3.6.1.0
Brotli==1.0.7
cachetools==3.1.1
celery==4.3.0
certifi==2019.9.11
cffi==1.12.2
chardet==3.0.4
confusable-homoglyphs==3.2.0
coverage==4.5.4
cryptography==2.6.1
csscompressor==0.9.5
cycler==0.10.0
defusedxml==0.5.0
dj-database-url==0.4.1
Django==2.1.5
django-admin==2.0.0
django-allauth==0.38.0
django-appconf==1.0.3
django-argonauts==1.2.0
django-autoslug==1.9.6
django-autoslugged==2.0.0
django-blog-zinnia==0.20
django-bootstrap-form==3.4
django-bootstrap4==1.0.1
django-ckeditor==5.8.0
django-colorful==1.3
django-compressor==2.3
django-contrib-comments==1.9.1
django-countries==5.3.3
django-countries-plus==1.2.1
django-crispy-forms==1.7.2
django-dashing==0.3.4
django-debug-toolbar==1.11
django-easy-subscription==0.1.0
django-el-pagination==3.2.4
django-excel-response2==2.0.8
django-extensions==2.2.5
django-filter==2.2.0
django-formtools==2.1
django-image-optimizer==1.0.0
django-image-renderer==0.1.3
django-imagekit==4.0.2
django-jet==1.0.8
django-js-asset==1.2.2
django-meta==1.5.2
django-mfa==2.0
django-modelcluster==4.4
django-mptt==0.10.0
django-oauth-toolkit==1.2.0
django-organizations==1.1.2
django-otp==0.7.3
django-otp-yubikey==0.5.2
django-phonenumber-field==1.3.0
django-picklefield==2.0
django-post-office==3.2.1
django-prometheus==1.1.0
django-recaptcha==2.0.5
django-registration==3.0.1
django-six==1.0.4
django-social-share==1.3.2
django-tagging==0.4.6
django-taggit==0.24.0
django-tinymce==2.8.0
django-treebeard==4.3
django-two-factor-auth==1.9.1
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
django-xmlrpc==0.1.8
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
draftjs-exporter==2.1.7
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
flake8==3.7.8
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
gunicorn==19.6.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.22
isort==4.3.4
joblib==0.14.0
jsonfield==2.0.2
kiwisolver==1.1.0
kombu==4.6.4
lazy-object-proxy==1.3.1
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
mots-vides==2015.5.11
nose==1.3.7
numpy==1.17.3
oauthlib==2.1.0
optimized-image==0.0.1
packaging==19.1
pandas==0.25.3
pep8==1.7.1
phonebook==1.0.0
phonenumbers==8.10.20
pilkit==2.0
Pillow==5.3.0
pinax-blog==7.0.4
pinax-images==3.0.1
pkg-resources==0.0.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
psycopg2==2.7
psycopg2-binary==2.8.3
puput==1.0.5
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.1
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.2.2
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.1
python-decouple==3.1
python-memcached==1.59
python-u2flib-server==5.0.0
python3-openid==3.1.0
pytz==2018.9
qrcode==6.1
rcssmin==1.0.6
regex==2019.11.1
requests==2.21.0
requests-oauthlib==1.0.0
rjsmin==1.1.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
screen==1.0.1
sentry-sdk==0.13.1
six==1.12.0
slimit==0.8.1
sqlparse==0.3.0
stripe==2.27.0
tinify==1.5.1
toml==0.10.0
tox==3.14.0
typed-ast==1.1.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
virtualenv==16.7.7
wagtail==2.6.2
wcwidth==0.1.7
webencodings==0.5.1
weboffice365==0.1
whitenoise==3.1
Willow==1.1
wrapt==1.10.11
xlwt==1.3.0
YubiOTP==0.2.2.post1
zipp==0.6.0
